+++ python/marvin/extern/__init__.py
@@ -9,14 +9,15 @@ def _import_module(name, relpath):
     """Imports a module from a path relative to this one's."""
 
     try:
-        return sys.modules[name]
-    except KeyError:
+        fp, pathname, description = imp.find_module(name)
+        return imp.load_module(name, fp, pathname, description)
+    except ImportError:
         pass
 
     try:
         path = os.path.join(os.path.dirname(__file__), relpath)
-        fp, filename, data = imp.find_module(name, [path])
-        return imp.load_module(name, fp, filename, data)
+        fp, filename, description = imp.find_module(name, [path])
+        return imp.load_module(name, fp, filename, description)
     except IOError:
         warnings.warn('Marvin cannot import {0}'.format(name), ImportWarning)
         return None